feat: use react-error-boundary
