+++ .storybook/preview.tsx
@@ -1,6 +1,7 @@
 import '@fontsource/ibm-plex-mono'
 import 'inter-ui/inter.css'
 
+import { ErrorBoundary } from 'react-error-boundary'
 import { withThemeFromJSXProvider } from '@storybook/addon-themes'
 import type { Args, Decorator, Preview } from '@storybook/react'
 import mockdate from 'mockdate'
@@ -16,7 +17,7 @@ import { theme } from '~/theme'
 
 import { Box, Skeleton } from '@chakra-ui/react'
 import { initialize, mswDecorator } from 'msw-storybook-addon'
-import ErrorBoundary from '~/components/ErrorBoundary'
+import {DefaultFallback } from '~/components/ErrorBoundary'
 import Suspense from '~/components/Suspense'
 import { format } from 'date-fns'
 import {
@@ -67,7 +68,8 @@ const SetupDecorator: Decorator = (story) => {
     })
   )
   return (
-    <ErrorBoundary>
+
+    <ErrorBoundary FallbackComponent={DefaultFallback}>
       <Suspense fallback={<Skeleton width="100vw" height="100vh" />}>
         <trpc.Provider client={trpcClient} queryClient={queryClient}>
           <QueryClientProvider client={queryClient}>
