+++ package.json
@@ -72,6 +72,7 @@
     "pino-pretty": "^10.2.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-error-boundary": "^4.0.12",
     "react-hook-form": "^7.45.2",
     "react-icons": "^4.8.0",
     "superjson": "^1.12.3",
