--- src/components/ErrorBoundary/ErrorBoundary.types.ts
@@ -1,11 +1,0 @@
-import { type ReactNode } from 'react'
-
-export type ErrorBoundaryState = {
-  hasError: boolean
-  error?: unknown
-}
-
-export type ErrorBoundaryProps = {
-  children: ReactNode
-  fallback?: ReactNode
-}
